+++ bevy_asset_loader/tests/continues_to_failure_state.rs
@@ -9,13 +9,12 @@ use bevy_asset_loader::loading_state::{LoadingState, LoadingStateAppExt};
 #[cfg(all(
     not(feature = "2d"),
     not(feature = "3d"),
-    not(feature = "progress_tracking"),
-    not(feature = "stageless")
+    not(feature = "progress_tracking")
 ))]
 #[test]
 fn continues_to_failure_state() {
     App::new()
-        .add_state(MyStates::Load)
+        .add_state::<MyStates>()
         .add_plugins(MinimalPlugins)
         .add_plugin(AssetPlugin::default())
         .add_loading_state(
@@ -50,8 +49,9 @@ struct Audio {
     no_loader_for_ogg_files: Handle<AudioSource>,
 }
 
-#[derive(Clone, Eq, PartialEq, Debug, Hash)]
+#[derive(Clone, Eq, PartialEq, Debug, Hash, Default, States)]
 enum MyStates {
+    #[default]
     Load,
     Error,
     Next,