+++ bevy_asset_loader/tests/continues_without_collection.rs
@@ -8,13 +8,12 @@ use bevy_asset_loader::loading_state::{LoadingState, LoadingStateAppExt};
 #[cfg(all(
     not(feature = "2d"),
     not(feature = "3d"),
-    not(feature = "progress_tracking"),
-    not(feature = "stageless")
+    not(feature = "progress_tracking")
 ))]
 #[test]
 fn continues_without_collection() {
     App::new()
-        .add_state(MyStates::Load)
+        .add_state::<MyStates>()
         .add_plugins(MinimalPlugins)
         .add_plugin(AssetPlugin::default())
         .add_loading_state(LoadingState::new(MyStates::Load).continue_to_state(MyStates::Next))
@@ -46,8 +45,9 @@ impl Default for TestState {
     }
 }
 
-#[derive(Clone, Eq, PartialEq, Debug, Hash)]
+#[derive(Clone, Eq, PartialEq, Debug, Hash, Default, States)]
 enum MyStates {
+    #[default]
     Load,
     Next,
 }