+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Option to load folders as maps of path to handle (resolves [#32](https://github.com/NiklasEi/bevy_asset_loader/issues/32))
 - Changed the default file ending for standard dynamic asset files from `.assets` to `.assets.ron`
 - Fix: multiple calls of `with_dynamic_collections` (resolves [#99](https://github.com/NiklasEi/bevy_asset_loader/issues/99))
+- Update to Bevy `0.10`
+  - Removed features `stageless` and `progress_tracking_stageless`
 
 ## v0.14.1
 - Make sure that the `RonAssetPlugin` for `StandardDynamicAssetCollection` is only added once